MON-6897 - Make Merlin binlog size configurable
shared.h - Added binlog_max_msize/fsize

shared.c - Added binlog_max_msize/fsize and increased the default value to
500/5000 MB.

node.c - Using the binlog_max_msize/fsize variable in node_binlog_add instead
of hardcoded values.

configuration.c - Added configuration validation for the new binlog options

bltest.c - Updated msize/fsize type from uint to unsigned long long int

HOWTO - Added some information about the new configuration option

+++ shared/binlog.h
@@ -47,7 +47,7 @@ extern const char *binlog_path(binlog *bl);
  * @param flags Decide what to do with an already existing file at path
  * @return A binlog object on success, NULL on errors.
  */
-extern binlog *binlog_create(const char *path, unsigned int msize, unsigned int fsize, int flags);
+extern binlog *binlog_create(const char *path, unsigned long long int msize, unsigned long long int fsize, int flags);
 
 /**
  * Get the number of unread entries in the binlog
